+++ client/src/input/Block.js
@@ -20,7 +20,8 @@ const _defaults = {
   enabled: true,
   collapsed: false,
   modified: true,
-  static: false
+  static: false,
+  showButtons: true
 }
 
 const _resources = {}
@@ -88,7 +89,8 @@ export default Garnish.Base.extend({
       collapsed: !!settings.collapsed,
       level: settings.level,
       modified: this._modified,
-      sortOrder: this._id
+      sortOrder: this._id,
+      showButtons: settings.showButtons
     }))
 
     NS.leave()