+++ client/src/input/Input.js
@@ -25,6 +25,7 @@ const _defaults = {
   inputId: null,
   maxBlocks: 0,
   maxTopBlocks: 0,
+  maxLevel: 0,
   static: false
 }
 
@@ -44,6 +45,7 @@ export default Garnish.Base.extend({
     this._name = settings.name
     this._maxBlocks = settings.maxBlocks
     this._maxTopBlocks = settings.maxTopBlocks
+    this._maxLevel = settings.maxLevel
     this._static = settings.static
     this._ownerId = null
 
@@ -153,6 +155,7 @@ export default Garnish.Base.extend({
         items: blockType.getChildBlockItems(this.getItems()),
         maxBlocks: this.getMaxBlocks()
       })
+      bInfo.showButtons = !this.atMaxLevel(bInfo.level)
 
       const block = new Block(bInfo)
       this.addBlock(block, -1, bInfo.level | 0, false)
@@ -420,6 +423,14 @@ export default Garnish.Base.extend({
 
   getMaxTopBlocks () {
     return this._maxTopBlocks
+  },
+
+  getMaxLevel () {
+    return this._maxLevel
+  },
+
+  atMaxLevel (level) {
+    return this._maxLevel && level + 1 >= this._maxLevel
   },
 
   getSelectedBlocks () {
@@ -709,7 +720,8 @@ export default Garnish.Base.extend({
             level: renderedBlock.level | 0,
             buttons: newButtons,
             enabled: !!renderedBlock.enabled,
-            collapsed: !!renderedBlock.collapsed
+            collapsed: !!renderedBlock.collapsed,
+            showButtons: !this.atMaxLevel(renderedBlock.level | 0)
           })
 
           newBlocks.push(newBlock)
@@ -754,7 +766,8 @@ export default Garnish.Base.extend({
       buttons: new Buttons({
         items: e.blockType.getChildBlockItems(this.getItems()),
         maxBlocks: this.getMaxBlocks()
-      })
+      }),
+      showButtons: !this.atMaxLevel(e.level)
     })
 
     this.addBlock(block, e.index, e.level)