+++ src/validators/FieldValidator.php
@@ -20,14 +20,29 @@ class FieldValidator extends Validator
     public $maxTopBlocks;
 
     /**
+     * @var int|null The maximum level blocks can be nested in the field.  If not set, there is no limit.
+     */
+    public $maxLevel;
+
+    /**
      * @var string|null A user-defined error message to be used if the field's `maxTopBlocks` is exceeded.
      */
     public $tooManyTopBlocks;
 
     /**
+     * @var string|null A user-defined error message to be used if the field's `maxLevel` is exceeded.
+     */
+    public $exceedsMaxLevel;
+
+    /**
      * @var string|null A user-defined error message to be used if a block type's `maxBlocks` is exceeded.
      */
     public $tooManyBlocksOfType;
+
+    /**
+     * @var array of Neo blocks
+     */
+    private $_blocks = [];
 
     /**
      * @inheritdoc
@@ -38,18 +53,10 @@ class FieldValidator extends Validator
 
         $field = $model->getFieldLayout()->getFieldByHandle(substr($attribute, 6));
         $value = $model->$attribute;
-        $blocks = $value->all();
+        $this->_blocks = $value->all();
 
-        // Check for max top-level blocks
-        if ($this->maxTopBlocks !== null) {
-            $topBlocks = array_filter($blocks, function($block) {
-                return (int)$block->level === 1;
-            });
-
-            if (count($topBlocks) > $this->maxTopBlocks) {
-                $this->addError($model, $attribute, $this->tooManyTopBlocks, ['maxTopBlocks' => $this->maxTopBlocks]);
-            }
-        }
+        $this->_checkMaxTopLevelBlocks($model, $attribute);
+        $this->_checkMaxLevel($model, $attribute);
 
         // Check for max blocks by block type
         $blockTypesCount = [];
@@ -60,7 +67,7 @@ class FieldValidator extends Validator
             $blockTypesById[$blockType->id] = $blockType;
         }
 
-        foreach ($blocks as $block) {
+        foreach ($this->_blocks as $block) {
             $blockTypesCount[$block->typeId] += 1;
         }
 
@@ -82,12 +89,50 @@ class FieldValidator extends Validator
     }
 
     /**
+     * Adds an error if the field exceeds its max top-level blocks.
+     */
+    private function _checkMaxTopLevelBlocks($model, $attribute)
+    {
+        if ($this->maxTopBlocks !== null) {
+            $topBlocks = array_filter($this->_blocks, function($block) {
+                return (int)$block->level === 1;
+            });
+
+            if (count($topBlocks) > $this->maxTopBlocks) {
+                $this->addError($model, $attribute, $this->tooManyTopBlocks, ['maxTopBlocks' => $this->maxTopBlocks]);
+            }
+        }
+    }
+
+    /**
+     * Adds an error if the field exceeds its max level.
+     */
+    private function _checkMaxLevel($model, $attribute)
+    {
+        $maxLevel = $this->maxLevel;
+
+        if ($maxLevel !== null) {
+            $tooHighBlocks = array_filter($this->_blocks, function($block) use($maxLevel) {
+                return ((int)$block->level) > $maxLevel;
+            });
+
+            if (!empty($tooHighBlocks)) {
+                $this->addError($model, $attribute, $this->exceedsMaxLevel, ['maxLevel' => $this->maxLevel]);
+            }
+        }
+    }
+
+    /**
      * Sets default error messages for any error messages that have not already been set.
      */
     private function _setDefaultErrorMessages()
     {
         if ($this->tooManyTopBlocks === null) {
             $this->tooManyTopBlocks = Craft::t('neo', '{attribute} should contain at most {maxTopBlocks, number} top-level {maxTopBlocks, plural, one{block} other{blocks}}.');
+        }
+
+        if ($this->exceedsMaxLevel === null) {
+            $this->exceedsMaxLevel = Craft::t('neo', '{attribute} has a max level of {maxLevel, number}, but has blocks exceeding that level.');
         }
 
         if ($this->tooManyBlocksOfType === null) {